requirements.txt
algoliasearch==3.0.0
algoliasearch-django==3.0.0
asgiref==3.7.2
bleach==6.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.33.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
Django==4.1.9
django-admin-volt==1.0.10
django-anymail==10.1
django-ckeditor==6.5.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.0.0
django-crontab==0.7.1
django-extensions==3.2.1
django-filter==23.2
django-js-asset==2.0.0
django-rest-swagger==2.2.0
django-solo==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.5
easy-thumbnails==2.8.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
phonenumbers==8.13.13
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pymemcache==4.0.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
webencodings==0.5.1
whitenoise==6.4.0
